forge fmt

+++ src/SimpleStorage.sol
@@ -2,24 +2,25 @@
 
 pragma solidity ^0.8.19; // solidity version
 
-
 contract SimpleStorage {
-    uint public myFavoriteNumber;
+    uint256 public myFavoriteNumber;
+
     // type Person
-    struct Person{
+    struct Person {
         uint256 favoriteNumber;
         string name;
     }
+
     Person[] public listOfPeople;
     // name -> favNum
-    mapping (string => uint256) public nameToFavoriteNumber;
+    mapping(string => uint256) public nameToFavoriteNumber;
 
     function store(uint256 _favoriteNumber) public virtual {
         myFavoriteNumber = _favoriteNumber;
     }
 
     // view, pure -> view means we are just reading state , pure means cant change AND also cant view
-    function retreive() public view returns(uint256){
+    function retreive() public view returns (uint256) {
         return myFavoriteNumber;
     }
 
@@ -28,6 +29,4 @@ contract SimpleStorage {
         listOfPeople.push(Person(_favoriteNumber, _name));
         nameToFavoriteNumber[_name] = _favoriteNumber;
     }
-
 }
-
